Temporary fix applied for CurrencyRateUpdateAbstract.php so that unit tests may complete correctly.

Conflicts:
	sugarcrm/clients/base/fields/enum/enum.js
	sugarcrm/sidecar
	sugarcrm/styleguide

+++ sugarcrm/tests/service/Bug51617Test.php
@@ -90,7 +90,9 @@ class Bug51617Test extends SOAPTestCase
     public function tearDown()
     {
         $this->df->deleteField($this->field);
-        $GLOBALS['db']->query("DELETE FROM accounts_cstm WHERE id_c = '{$this->_account->id}'");
+        if ($GLOBALS['db']->tableExists('accounts_cstm')) {
+            $GLOBALS['db']->query("DELETE FROM accounts_cstm WHERE id_c = '{$this->_account->id}'");
+        }
 
         SugarTestAccountUtilities::removeAllCreatedAccounts();
         SugarTestUserUtilities::removeAllCreatedAnonymousUsers();
